Break out Nonce
